product_brand_tag: add mixin
