+++ product_brand_tag/models/__init__.py
@@ -1,2 +1,3 @@
 from . import product_brand
+from . import product_brand_tag_mixin
 from . import product_brand_tag
